requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
asyncpg==0.29.0
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose==3.3.0
bcrypt==4.2.0
python-dotenv==1.0.1
email-validator==2.2.0
fastapi-mail==1.4.1
pytest==8.3.3
httpx==0.27.2
pytest-asyncio==0.24.0
psycopg[binary]==3.2.3

